config/errors: move under 'shared'
Since more than just errors are shared between multiple config versions,
let's collect all the things under 'shared'
