+++ config/util/parsingErrors.go
@@ -18,7 +18,7 @@ import (
 	"bytes"
 	"errors"
 
-	configErrors "github.com/coreos/ignition/config/errors"
+	configErrors "github.com/coreos/ignition/config/shared/errors"
 	"github.com/coreos/ignition/config/v2_3_experimental/types"
 	"github.com/coreos/ignition/config/validate/report"
