+++ config/v1/config.go
@@ -15,7 +15,7 @@
 package v1
 
 import (
-	"github.com/coreos/ignition/config/errors"
+	"github.com/coreos/ignition/config/shared/errors"
 	"github.com/coreos/ignition/config/util"
 	"github.com/coreos/ignition/config/v1/types"
 	"github.com/coreos/ignition/config/validate"
