+++ config/v2_1/types/file.go
@@ -17,7 +17,7 @@ package types
 import (
 	"fmt"
 
-	"github.com/coreos/ignition/config/errors"
+	"github.com/coreos/ignition/config/shared/errors"
 	"github.com/coreos/ignition/config/validate/report"
 )
