+++ config/v2_2/config.go
@@ -15,7 +15,7 @@
 package v2_2
 
 import (
-	"github.com/coreos/ignition/config/errors"
+	"github.com/coreos/ignition/config/shared/errors"
 	"github.com/coreos/ignition/config/v2_1"
 	"github.com/coreos/ignition/config/v2_2/types"
 	"github.com/coreos/ignition/config/validate"
