+++ config/v2_3_experimental/config.go
@@ -15,7 +15,7 @@
 package v2_3_experimental
 
 import (
-	"github.com/coreos/ignition/config/errors"
+	"github.com/coreos/ignition/config/shared/errors"
 	"github.com/coreos/ignition/config/v2_2"
 	"github.com/coreos/ignition/config/v2_3_experimental/types"
 	"github.com/coreos/ignition/config/validate"
